requirements.txt
Flask>=2.0.0
requests>=2.0.0
gunicorn>=20.0.0
Flask-OAuthlib==0.9.6
Flask-SQLAlchemy==2.5.1
pymysql==1.0.2
authlib==0.15.4
mysql-connector-python
stravalib